refactor  register
